Renaming Cli -> Analyzer
